Data-Persistence
Using SQL and Php to populate, manage, retrieve and update data

+++ Create_student_tbl.sql
@@ -2,9 +2,9 @@
  CREATE TABLE student(
   Firstname       VARCHAR(20)     NOT NULL,
   Surname         VARCHAR(20)     NOT NULL,
-  Algebra         DECIMAL(10,2)   NOT NULL,
-  Calculus        DECIMAL(10,2)   NOT NULL,
-  Programming     DECIMAL(10,2)   NOT NULL,
-  Databasess      DECIMAL(10,2)   NOT NULL,
+  Algebra         INT(10)		  NOT NULL,
+  Calculus        INT(10)	      NOT NULL,
+  Programming     INT(10)	      NOT NULL,
+  Databasess      INT(10)	      NOT NULL,
   PRIMARY KEY (Firstname)
 );
